requirements.txt
numpy
scikit-image
trimesh
pyrender
PyOpenGL==3.1.5
videoio
scipy
tqdm
